pyproject.toml
[project]
name = "celebrity-detector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "groq>=0.36.0",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain-groq>=1.1.0",
    "numpy>=2.3.5",
    "opencv-python>=4.11.0.86",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "ruff>=0.14.6",
]
